fix(framework): Wait before Control API RPCs
