+++ framework/py/flwr/cli/log.py
@@ -32,7 +32,11 @@ from flwr.common.logger import log as logger
 from flwr.proto.control_pb2 import StreamLogsRequest  # pylint: disable=E0611
 from flwr.proto.control_pb2_grpc import ControlStub
 
-from .utils import flwr_cli_grpc_exc_handler, init_channel_from_connection
+from .utils import (
+    flwr_cli_grpc_exc_handler,
+    init_channel_from_connection,
+    wait_for_control_api_channel,
+)
 
 
 class AllLogsRetrieved(BaseException):
@@ -218,6 +222,7 @@ def _log_with_control_api(
         If True, stream logs continuously; if False, print once.
     """
     channel = init_channel_from_connection(superlink_connection)
+    wait_for_control_api_channel(channel)
 
     if stream:
         start_stream(run_id, channel, CONN_REFRESH_PERIOD)
