+++ framework/py/flwr/cli/run/run.py
@@ -16,12 +16,10 @@
 
 
 import hashlib
-import time
 from pathlib import Path
 from typing import Annotated, Any, Literal
 
 import click
-import grpc
 import typer
 
 from flwr.cli.build import build_fab_from_disk, get_fab_filename
@@ -50,15 +48,10 @@ from ..utils import (
     flwr_cli_grpc_exc_handler,
     init_channel_from_connection,
     print_json_to_stdout,
+    wait_for_control_api_channel,
 )
 
 CONN_REFRESH_PERIOD = 60  # Connection refresh period for log streaming (seconds)
-SUPERLINK_UNAVAILABLE_MESSAGE = (
-    "Connection to the SuperLink is unavailable. Please check your network "
-    "connection and 'address' in the SuperLink connection configuration."
-)
-CONTROL_API_READY_TIMEOUT_SECONDS = 30
-CONTROL_API_READY_CHECK_INTERVAL_SECONDS = 1
 
 
 # pylint: disable-next=too-many-locals, too-many-branches, R0913, R0917
@@ -209,7 +202,7 @@ def _run_with_control_api(
             app_spec=app_spec or "",
         )
         with flwr_cli_grpc_exc_handler():
-            _wait_for_control_api_channel(channel)
+            wait_for_control_api_channel(channel)
             res = stub.StartRun(req)
 
         if res.HasField("note"):
@@ -249,25 +242,6 @@ def _run_with_control_api(
     finally:
         if channel:
             channel.close()
-
-
-def _wait_for_control_api_channel(
-    channel: grpc.Channel,
-    timeout: float = CONTROL_API_READY_TIMEOUT_SECONDS,
-    check_interval: float = CONTROL_API_READY_CHECK_INTERVAL_SECONDS,
-) -> None:
-    """Wait for the Control API channel to become ready before submitting a run."""
-    deadline = time.monotonic() + timeout
-    future = grpc.channel_ready_future(channel)
-    while True:
-        remaining = deadline - time.monotonic()
-        if remaining <= 0:
-            raise click.ClickException(SUPERLINK_UNAVAILABLE_MESSAGE)
-        try:
-            future.result(timeout=min(check_interval, remaining))
-            return
-        except grpc.FutureTimeoutError:
-            continue
 
 
 def _parse_federation_config_overrides(
